+++ RELEASE-NOTES.md
@@ -7,6 +7,7 @@ Latest release:
 
 * Dropped support for PHP 5.x
 * Dropped class aliases deprecated since Diff 1.0
+* Removed ListDiff and MapDiff, deprecated since Diff 0.5
 
 ## Version 2.1 (2016-09-01)